+++ packages/core/src/data/ome_zarr/0.4/bf2raw.ts
