+++ packages/core/src/data/ome_zarr/0.5/_version.ts
