+++ packages/core/src/data/ome_zarr/codegen.sh
