+++ packages/core/src/data/ome_zarr/resolve_schema.cjs
